[#02][be][user]style: 이메일 로그, 알림 로그 패키지 재설정
