--- back/src/main/java/com/rouby/notification/domain/entity/notificationlog/SendStatus.java
@@ -1,7 +1,0 @@
-package com.rouby.notification.domain.entity.notificationlog;
-
-public enum SendStatus {
-  SENT,
-  FAILED
-  ;
-}
