+++ back/src/main/java/com/rouby/notification/notificationlog/domain/entity/NotificationMessage.java
@@ -1,4 +1,4 @@
-package com.rouby.notification.domain.entity.notificationlog;
+package com.rouby.notification.notificationlog.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
