+++ back/src/main/java/com/rouby/notification/notificationlog/domain/entity/NotificationType.java
@@ -1,7 +1,6 @@
-package com.rouby.notificationtemplate.domain.entity;
+package com.rouby.notification.notificationlog.domain.entity;
 
 public enum NotificationType {
-
   SCHEDULE,
   ROUTINE,
   BRIEFING,
