+++ back/src/main/java/com/rouby/notification/notificationlog/domain/entity/SendStatus.java
@@ -1,0 +1,7 @@
+package com.rouby.notification.notificationlog.domain.entity;
+
+public enum SendStatus {
+  SENT,
+  FAILED
+  ;
+}
