+++ back/src/main/java/com/rouby/notification/notificationtemplate/domain/repository/NotificationTemplateRepository.java
@@ -1,0 +1,5 @@
+package com.rouby.notification.notificationtemplate.domain.repository;
+
+public interface NotificationTemplateRepository {
+
+}
