+++ back/src/main/java/com/rouby/notification/notificationtemplate/infrastructure/persistence/jpa/NotificationTemplateJpaRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.rouby.notification.notificationtemplate.infrastructure.persistence.jpa;
+
+public interface NotificationTemplateJpaRepositoryCustom {
+
+}
