--- back/src/main/java/com/rouby/notificationtemplate/domain/repository/NotificationTemplateRepository.java
@@ -1,5 +1,0 @@
-package com.rouby.notificationtemplate.domain.repository;
-
-public interface NotificationTemplateRepository {
-
-}
