--- back/src/main/java/com/rouby/notificationtemplate/infrastructure/persistence/jpa/NotificationTemplateJpaRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.rouby.notificationtemplate.infrastructure.persistence.jpa;
-
-public interface NotificationTemplateJpaRepositoryCustom {
-
-}
